+++ src/artifact-sharing/components/artifact-text.tsx
@@ -1,0 +1,9 @@
+import { Artifact, IArtifactSharingURL } from '../tokens';
+
+
+export namespace ArtifactText {
+  export interface IProps {
+    urlFactory: IArtifactSharingURL;
+    artifact?: Artifact;
+  }
+}
