+++ src/artifact-sharing/components/edit-artifact/edit-artifact-text.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { ArtifactText } from "../artifact-text";
+
+
+export class EditArtifactText extends React.Component<ArtifactText.IProps> {
+  render() {
+    return (
+      <div className='chi-ArtifactForm-text'>
+        <h2>Edit artifact</h2>
+      </div>
+    );
+  }
+}
